Improve further FE - endpoint for updating of the ship info, Move query from shipService to shipRepository, Add about app info

+++ src/main/java/com/greenfox/mockexam/controllers/PlanetController.java
@@ -17,7 +17,7 @@ public class PlanetController {
     }
 
     @DeleteMapping("/planets/{id}")
-    public String deletePlanet(@PathVariable short id) {
+    public String deletePlanet(@PathVariable long id) {
         shipService.deleteSomePlanet(id);
 
         return "redirect:/";
